refactor: cleaner file structure and better service-worker/popup usage
